fix system types

+++ abis/hades/nlink_t.h
@@ -2,7 +2,7 @@
 #ifndef _ABIBITS_NLINK_T_H
 #define _ABIBITS_NLINK_T_H
 
-typedef long nlink_t;
+typedef int nlink_t;
 
 #endif // _ABIBITS_NLINK_T_H
